% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/linkage.R
\name{createPlotDF}
\alias{createPlotDF}
\title{Compile plotting Data}
\usage{
createPlotDF(obsDF)
}
\value{
Runs the ShinyThings app
}
\description{
Runs a SensorThings API data visualisation tool
}
\examples{

shinyThingsApp()
}
