% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shinyGraphs.R
\name{createShinyGraph}
\alias{createShinyGraph}
\title{Create interactive plot for SensorThings data}
\usage{
createShinyGraph(obsDf)
}
\arguments{
\item{locDf}{Data frame with class "thingLocations"}
}
\value{
A list object containing information from url/Things
}
\description{
Creates interactive graph for display on the ShinyThings app
}
\examples{

}
