% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/locations.R
\name{defineThingLocation}
\alias{defineThingLocation}
\title{Assign "thingLocation" class to data.frame object}
\usage{
defineThingLocation(df)
}
\arguments{
\item{url}{The SensorThings API url containing the data in SensorThings web standard in string(!) format}
}
\value{
A list object containing information from url/Things
}
\examples{
x = senseLocations(https://toronto-bike-snapshot.sensorup.com/v1.0/)
x

}
