% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/leaflet.R
\name{expressMapFoI}
\alias{expressMapFoI}
\title{Directly load thing FoIs to leaflet map}
\usage{
expressMapFoI(url)
}
\arguments{
\item{url}{Path or URL to input file}
}
\description{
Uses senseFoI(), makeThingFoI() and mapThingFoI to build leaflet map directly
}
\examples{
leaf1 = expressMapFoI("https://tasking-test.sensorup.com/v1.0")
leaf1

leaf2 = expressMapFoI("https://toronto-bike-snapshot.sensorup.com/v1.0")
leaf2
}
