% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/leaflet.R
\name{expressMapThings}
\alias{expressMapThings}
\title{Directly load thing locations to leaflet map}
\usage{
expressMapThings(url)
}
\arguments{
\item{url}{Path or URL to input file}
}
\value{
Object of class "track"
}
\description{
Uses senseLocations(), makeThingLocation() and mapThings
}
\examples{
map1 = expressMapThings("https://tasking-test.sensorup.com/v1.0/")
map1

map2 = expressMapThings("https://toronto-bike-snapshot.sensorup.com/v1.0/")
map2
}
