% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/linkage.R
\name{getObservations}
\alias{getObservations}
\title{Links features of interest with classes necessary for plotting}
\usage{
getObservations(url, FoIID)
}
\value{
Runs the ShinyThings app
}
\description{
Runs a SensorThings API data visualisation tool
}
\examples{

shinyThingsApp()
}
