% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/observations.R
\name{getStreamObservations}
\alias{getStreamObservations}
\title{Load datastreams and observations into R}
\usage{
getStreamObservations(url, streamID)
}
\arguments{
\item{url}{A SensorThings API base url (string!), a thing Id}
}
\value{
A "streamObject" dataframe containing data from url/Thing(id)/Datastreams
}
\description{
This function parses Thing SensorThings JSON data and stores it in an R data frame. Contains the complete Thing data for further processing.
}
