% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/observations.R
\name{getThingDatastreams}
\alias{getThingDatastreams}
\title{Lists the datastreams of a thing and loads them into R}
\usage{
getThingDatastreams(url, thingID)
}
\arguments{
\item{url}{A SensorThings API base url (string!), a thing Id}
}
\value{
A "streamObject" dataframe containing data from url/Thing(id)/Datastreams
}
\description{
This function parses Thing SensorThings JSON data and stores it in an R data frame. Contains the complete Thing data for further processing.
}
