% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/foi.R
\name{makeThingFoI}
\alias{makeThingFoI}
\title{Create "thingFoI" data frame}
\usage{
makeThingFoI(FoIDF)
}
\arguments{
\item{FoIDF}{Dataframe of class "foiObject"}
}
\value{
Dataframe of class "thingFoI" and "mapThing"
}
\description{
Creates a data frame from "foiObject" data with the added classes "thingFoI" and "mapThing". "thingFoI" objects contain web links to the observations of every feature of interest, granting access to the observed data. "mapThing" objects can be imported onto a leaflet map.
}
\examples{
n = senseFoI("http://example.sensorup.com/v1.0")
u = makeThingFoI(n)
head(u)
}
