% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/locations.R
\name{makeThingLocation}
\alias{makeThingLocation}
\title{Assign "thingLocation" class to data frame object}
\usage{
makeThingLocation(locationDF)
}
\arguments{
\item{locationDF}{A data frame created using senseLocations (formatted according to SensorThings API)}

\item{df}{A data frame created using senseLocations (formatted according to SensorThings API)}
}
\value{
The input data frame with added class "thingLocation"

Data frame of class "thingLocation"
}
\description{
Appends the "thingLocation" class to the already present object classes

Create a "thingLocation" data frame from a previously parsed SensorThings JSON location object
}
\examples{
x = senseLocations(https://toronto-bike-snapshot.sensorup.com/v1.0/)
y = defineThingLocation(x)
y
n = senseLocations("http://example.sensorup.com/v1.0/")
u = makeThingLocation(n)
u

}
