% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/leaflet.R
\name{mapThingFoI}
\alias{mapThingFoI}
\title{Display SensorThing Features of Interest on a map}
\usage{
mapThingFoI(foiDf)
}
\arguments{
\item{foiDf}{Data frame with class "mapThing"}
}
\value{
A list object containing information from url/FeaturesOfInterest
}
\description{
Creates markers from "mapThing" objects and loads them onto a Leaflet map
}
\examples{
x = senseFoI("https://toronto-bike-snapshot.sensorup.com/v1.0")
y = makeThingFoI(x)
mapThingFoI(y)
}
