% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/leaflet.R
\name{mapThingLocations}
\alias{mapThingLocations}
\title{Display SensorThing locations on a map}
\usage{
mapThingLocations(locDf)
}
\arguments{
\item{locDf}{Data frame with class "thingLocations" and "mapThing"}
}
\value{
A list object containing information from url/Things
}
\description{
Creates markers from "thingLocation" objects and loads them onto a Leaflet map
}
\examples{
x = senseLocations("https://toronto-bike-snapshot.sensorup.com/v1.0")
y = makeThingLocation(x)
mapThingLocations(y)
}
