% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/leaflet.R
\name{mapThings}
\alias{mapThings}
\title{Loads location data from any SensorThings API}
\usage{
mapThings(locDf)
}
\arguments{
\item{url}{The SensorThings API url containing the data in SensorThings web standard in string(!) format}
}
\value{
A list object containing information from url/Things
}
\examples{
x = createThingMap("https://toronto-bike-snapshot.sensorup.com/v1.0/")
y = locToDf(x)
locationLeaflet
}
