% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datastreams.R
\name{senseDataStreams}
\alias{senseDataStreams}
\title{Loads datastream data from any SensorThings API}
\usage{
senseDataStreams(url)
}
\arguments{
\item{url}{A SensorThings API base url (string!) containing the data in SensorThings web standard}
}
\value{
A data frame object containing data from url/Datastreams
}
\description{
This function parses SensorThings JSON data and stores it in an R data frame
}
\examples{
ds = senseDataStreams("https://toronto-bike-snapshot.sensorup.com/v1.0")
ds
}
