% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/foi.R
\name{senseFoI}
\alias{senseFoI}
\title{Load FeatureOfInterest data from SensorThings APIs}
\usage{
senseFoI(url)
}
\arguments{
\item{url}{A SensorThings API base url (string!)}
}
\value{
A "foiObject" dataframe containing data from url/FeaturesOfInterest
}
\description{
This function parses FoI SensorThings JSON data and stores it in an R data frame. Contains the complete FoI data for further processing.
}
\examples{
x = senseFoI("https://toronto-bike-snapshot.sensorup.com/v1.0")
head(x)

\dontrun{
v = senseFoI("https://tasking-test.sensorup.com/v1.0")
v

a = senseFoI("http://example.sensorup.com/v1.0")
a
}
}
