% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/things.R
\name{senseThings}
\alias{senseThings}
\title{Load Thing data from SensorThings APIs}
\usage{
senseThings(url)
}
\arguments{
\item{url}{A SensorThings API base url (string!)}
}
\value{
A "thingObject" dataframe containing data from url/Things
}
\description{
This function parses Thing SensorThings JSON data and stores it in an R data frame. Contains the complete Thing data for further processing.
}
\examples{
s1 = senseThings("https://toronto-bike-snapshot.sensorup.com/v1.0")
head(s1)

\dontrun{
s2 = senseThings("https://tasking-test.sensorup.com/v1.0")
s2

s3 = senseThings("http://example.sensorup.com/v1.0")
class(s3)
}
}
