% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shinyThings.R
\name{shinyThings}
\alias{shinyThings}
\title{Shiny App for SensorThings API}
\usage{
shinyThings()
}
\value{
Runs the ShinyThings app
}
\description{
Runs a SensorThings API data visualisation tool
}
\examples{

shinyThingsApp()
}
