% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shinyThings.R
\name{shinyThingsApp}
\alias{shinyThingsApp}
\title{Shiny App for SensorThings API}
\usage{
shinyThingsApp()
}
\value{
Runs the ShinyThings app
}
\description{
Runs a SensorThings API data visualisation tool
}
\examples{

shinyThingsApp()
}
