% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/leaflet.R
\name{thingsToMap}
\alias{thingsToMap}
\title{Directly plot thing locations to map}
\usage{
thingsToMap(url)
}
\arguments{
\item{url}{Path or URL to input file}
}
\value{
Object of class "track"
}
\description{
Uses
}
\examples{
map1 = thingsToMap("https://tasking-test.sensorup.com/v1.0/")
map1

map2 = thingsToMap("https://toronto-bike-snapshot.sensorup.com/v1.0/")
map2
}
