% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/leaflet.R
\name{expressMapLocations}
\alias{expressMapLocations}
\title{Directly load thing locations to leaflet map}
\usage{
expressMapLocations(url)
}
\arguments{
\item{url}{Path or URL to input file}
}
\description{
Uses senseLocations(), makeThingLocation() and mapThings to build leaflet map with a single function
}
\examples{
map1 = expressMapLocations("https://tasking-test.sensorup.com/v1.0")
map1

\dontrun{
map2 = expressMapLocations("https://toronto-bike-snapshot.sensorup.com/v1.0")
map2
}
}
