% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/locations.R
\name{getLocationThings}
\alias{getLocationThings}
\title{Gets the things (e.g. abstraction for data generating devices) at a location with specific ID}
\usage{
getLocationThings(url, locId)
}
\arguments{
\item{url}{A SensorThings API url (string!) containing the data in SensorThings web standard}

\item{locId}{The location Id of a location within the sensor network (must be present within the SensorThings network of the input url!)}
}
\value{
A data frame object containing data from url/Location/Things
}
\description{
Gets the things (e.g. abstraction for data generating devices) at a location with specific ID and stores their information in an R data frame
}
\examples{

things = getLocationThings("https://toronto-bike-snapshot.sensorup.com/v1.0", 1462)
things
}
