% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/locations.R
\name{locToDf}
\alias{locToDf}
\title{Create Track Object}
\usage{
locToDf(locationDF)
}
\arguments{
\item{input}{Path or URL to input file}

\item{headerBool}{Boolean value for presence of header in the file}

\item{sepStr}{String of the separator character (e.g. ",")}
}
\value{
Object of class "track"
}
\description{
Reads table format .txt files
}
\examples{
dataPath <- system.file("extdata", "data_pm10.csv", package="tracks4r")

x = readTrack(dataPath)

class(x)


}
