% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/foi.R
\name{makeThingFoI}
\alias{makeThingFoI}
\title{Create a "thingFoI" data frame}
\usage{
makeThingFoI(FoIDF)
}
\arguments{
\item{FoIDF}{A data frame created using senseFoI (formatted according to SensorThings API)}
}
\value{
Data frame of class "thingFoI"
}
\description{
Creates a data frame from a previously parsed SensorThings JSON FoI object with the added class "thingFoI"
}
\examples{
n = senseFoI("http://example.sensorup.com/v1.0")
u = makeThingFoI(n)
head(u)
}
