% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FoI.R
\name{makeThingFoI}
\alias{makeThingFoI}
\title{Create a "thingLocation" data frame}
\usage{
makeThingFoI(FoIDF)
}
\arguments{
\item{FoIDF}{A data frame created using senseLocations (formatted according to SensorThings API)}
}
\value{
Data frame of class "thingLocation"
}
\description{
Creates a data frame from a previously parsed SensorThings JSON location object with the added class "thingLocation"
}
\examples{
n = senseLocations("http://example.sensorup.com/v1.0")
u = makeThingLocation(n)
u

}
