% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/locations.R
\name{makeThingLocation}
\alias{makeThingLocation}
\title{create a "thingLocation" object from}
\usage{
makeThingLocation(locationDF)
}
\arguments{
\item{locationDF}{Path or URL to input file}
}
\value{
Object of class "track"
}
\description{
Reads table format .txt files
}
\examples{
n = senseLocations("http://example.sensorup.com/v1.0/")
u = locToDf(n)
u

}
