% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/leaflet.R
\name{mapThings}
\alias{mapThings}
\title{Loads location data from any SensorThings API}
\usage{
mapThings(locDf)
}
\arguments{
\item{locDf}{Data frame with class "thingLocations"}
}
\value{
A list object containing information from url/Things
}
\examples{
x = senseLocations("https://toronto-bike-snapshot.sensorup.com/v1.0/")
y = makeThingLocation(x)
mapThings(y)
}
