% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/foi.R
\name{senseFoI}
\alias{senseFoI}
\title{Loads FoI data from any SensorThings API}
\usage{
senseFoI(url)
}
\arguments{
\item{url}{A SensorThings API url (string!) containing the data in SensorThings web standard}
}
\value{
A data frame object containing data from url/FeaturesOfInterest
}
\description{
This function parses SensorThings JSON data and stores it in an R data frame
}
\examples{
x = senseFoI("https://toronto-bike-snapshot.sensorup.com/v1.0")
x

\dontrun{
v = senseFoI("https://tasking-test.sensorup.com/v1.0")
v

a = senseFoI("http://example.sensorup.com/v1.0")
a
}
}
