% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/locations.R
\name{senseLocations}
\alias{senseLocations}
\title{Loads location data from any SensorThings API}
\usage{
senseLocations(url)
}
\arguments{
\item{url}{The SensorThings API url containing the data in SensorThings web standard in string(!) format}
}
\value{
A data frame object containing data from url/Things
}
\examples{
x = senseLocations("https://toronto-bike-snapshot.sensorup.com/v1.0/")
x

v = senseLocations("https://tasking-test.sensorup.com/v1.0/")
v

a = senseLocations("http://example.sensorup.com/v1.0/")
a
}
