% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/locations.R
\name{senseLocations}
\alias{senseLocations}
\title{Loads location data from any SensorThings API}
\usage{
senseLocations(url)
}
\arguments{
\item{url}{A SensorThings API url (string!) containing the data in SensorThings web standard}
}
\value{
A data frame object containing data from url/Things
}
\description{
This function parses SensorTHings JSON data and stores it in an R data frame
}
\examples{
x = senseLocations("https://toronto-bike-snapshot.sensorup.com/v1.0")
x

v = senseLocations("https://tasking-test.sensorup.com/v1.0")
v

a = senseLocations("http://example.sensorup.com/v1.0")
a
}
