% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/things.R
\name{senseThings}
\alias{senseThings}
\title{Loads location data from any SensorThings API}
\usage{
senseThings(url)
}
\arguments{
\item{url}{A SensorThings API url (string!) containing the data in SensorThings web standard}
}
\value{
A data frame object containing data from url/Things
}
\description{
This function parses SensorTHings JSON data and stores it in an R data frame
}
\examples{
s1 = senseThings("https://toronto-bike-snapshot.sensorup.com/v1.0")
head(s1)

\dontrun{
s2 = senseThings("https://tasking-test.sensorup.com/v1.0")
s2

s3 = senseThings("http://example.sensorup.com/v1.0")
class(s3)
}
}
