% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shinyThings.R
\name{shinyThings}
\alias{shinyThings}
\title{Shiny App for SensorThings API}
\usage{
shinyThings()
}
\value{
Shiny app UI
}
\description{
Runs a SensorThings API data visualisation tool. Loads all station Locations (url/Locations) on a map when given an input SensorThings URL. Associated Thing info is displayed when map markers are clicked.
}
\examples{
\dontrun{
shinyThings()
}
}
