% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/locations.R
\name{thingsToMap}
\alias{thingsToMap}
\title{Directly plot thing locations to map}
\usage{
thingsToMap(url)
}
\arguments{
\item{url}{Path or URL to input file}
}
\value{
Object of class "track"
}
\description{
Reads table format .txt files
}
\examples{
dataPath <- system.file("extdata", "data_pm10.csv", package="tracks4r")

x = readTrack(dataPath)

class(x)

}
